# FilesService/swagger.yaml
basePath: /files/
definitions:
...
    empty:
        type: object
        x-go-name: EmptyResponse
        x-go-package: github.com/KrzysztofSieczkiewicz/go--model-viewer-backend/FilesService/response
    fileByteStream:
        properties:
            File:
                description: File bytestream
                items:
                    format: uint8
                    type: integer
                type: array
        type: object
        x-go-name: FileResponse
        x-go-package: github.com/KrzysztofSieczkiewicz/go--model-viewer-backend/FilesService/response
    fileUrl:
        properties:
            filename:
                description: filename
                type: string
                x-go-name: Filename
            url:
                description: url pointing to the resource
                type: string
                x-go-name: URL
        type: object
        x-go-name: FileUrlResponse
        x-go-package: github.com/KrzysztofSieczkiewicz/go--model-viewer-backend/FilesService/response
...
    message:
        properties:
            message:
                type: string
                x-go-name: Message
        type: object
        x-go-name: MessageResponse
        x-go-package: github.com/KrzysztofSieczkiewicz/go--model-viewer-backend/FilesService/response
host: localhost:9090
info:
    description: Documentation of FilesApi
    title: FilesService.
    version: 1.0.0
paths:
...
    /{id}&{expires}&{signature}:
        get:
            description: Return an image from imageset. Can only be accessed by signed URLs
            operationId: getImage
            produces:
                - application/octet-stream
                - application/json
            responses:
                "200":
                    description: fileByteStream
                    schema:
                        $ref: '#/definitions/fileByteStream'
                "400":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "403":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "404":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "500":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
            tags:
                - images
...
    /images/{category}/{id}:
        delete:
            consumes:
                - application/json
            description: Remove image from the image set
            operationId: deleteImage
            produces:
                - application/json
            responses:
                "204":
                    description: empty
                    schema:
                        $ref: '#/definitions/empty'
                "404":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "500":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
            tags:
                - images
        get:
            consumes:
                - application/json
            description: Return a signed url to requested image
            operationId: getImageUrl
            produces:
                - application/json
            responses:
                "200":
                    description: fileUrl
                    schema:
                        $ref: '#/definitions/fileUrl'
                "404":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "500":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
            tags:
                - images
        post:
            consumes:
                - multipart/form-data
            description: Add an image to the existing set
            operationId: postImage
            produces:
                - application/json
            responses:
                "201":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "400":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "403":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "500":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
            tags:
                - images
        put:
            consumes:
                - multipart/form-data
            description: Update an image in the image set
            operationId: putImage
            produces:
                - application/json
            responses:
                "200":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "404":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "500":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
            tags:
                - images
schemes:
    - http
    - https
swagger: "2.0"